+++ frontend/src/app/pages/top/index.tsx
@@ -4,13 +4,21 @@ import { Page } from "./page";
 
 export const Top: React.FC = () => {
   return (
-    <Page contest={contest}/>
+    <Page contests={contest}/>
   );
 }
 
-const contest = {
-  category: "NBA（バスケ）",
-  title: "Los Angels Lakers vs Golden State Warriors",
-  startAt: "2020.06.01｜11:00",
-  status: "upcoming" as const
-}
+const contest = [
+  {
+    category: "NBA（バスケ）",
+    title: "Los Angels Lakers vs Golden State Warriors",
+    startAt: "2020.06.01｜11:00",
+    status: "upcoming" as const
+  },
+  {
+    category: "NBA（バスケ）",
+    title: "Los Angels Lakers vs Golden State Warriors",
+    startAt: "2020.06.01｜11:00",
+    status: "open" as const
+  },
+];
